requirements.txt
Flask
Flask-Bootstrap
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
scikit-image
opencv-python
pyAudioAnalysis


